cluster: never send empty alert batches
If the configuration changes while a send_batch task is waiting, and no
new edges are added, the previous logic would send an empty batch.

This change prevents that by checking for an empty batch before sending.

Also replaces the task's started_at with a bool because it's never read.