cluster: fix rejoin bug exposed in test
Prior to this change, a node rejoining while its hostname is still
in the cluster would not update its metadata, because while joining
it would respond to probes (preventing the old version from being
kicked).

This modifies the probe rpc to fail with status unavailable if the
node isn't a member of any configuration, so any node which is trying
to join a cluster won't be deadlocked by maintaining the old entry.

This also modifies the join protocol so that a duplicate host will
not be allowed to join (forcing a proposal), instead of forwarding
the original host's observer information.